Major moving Worker-scoped globals from Window to WindowBase
